% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.PSM.R
\name{coef.PSM}
\alias{coef.PSM}
\title{Extract Model Coefficients}
\usage{
\method{coef}{PSM}(object)
}
\arguments{
\item{object}{an object of class 'PSM' returned by PartialSplines function}
}
\value{
a list containing the coefficients
}
\description{
Extracts model coefficients beta and delta
}
\examples{

x <- data.frame(x = rnorm(100))
tx <- rnorm(100)
eps <- rnorm(100)
y <- x[,1] + tx^4 + eps
psm <- PartialSplines(y, x, tx)
coef(psm)

}
