% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.PSM.R
\name{fitted.PSM}
\alias{fitted.PSM}
\title{Extract Model Fitted Values}
\usage{
\method{fitted}{PSM}(object)
}
\arguments{
\item{object}{an object of class 'PSM'}
}
\value{
returns a vector containing fitted values
}
\description{
extracts fitted values from a Partial Splines Model
}
\examples{
x <- data.frame(x = rnorm(100))
tx <- rnorm(100)
eps <- rnorm(100)
y <- x[,1] + tx^4 + eps
psm <- PartialSplines(y, x, tx)
fitted(psm)
}
