% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialSplines.control.R
\name{PartialSplines.control}
\alias{PartialSplines.control}
\title{Auxiliary for control PartialSplines fitting}
\usage{
PartialSplines.control(
  y,
  x,
  tx,
  prop.knots,
  intercept,
  n.basis,
  alpha,
  tol,
  b.start,
  d.start
)
}
\value{

}
\description{
It returns an error if at least one condition is not satisfied, it has an internal use inside PartialSplines function
}
\examples{
## do not use

}
