% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialSplines.R
\name{PartialSplines}
\alias{PartialSplines}
\title{Estimates a partial splines model}
\usage{
PartialSplines(
  y,
  x,
  tx,
  prop.knots = NULL,
  intercept = T,
  n.basis = 15,
  alpha = NULL,
  tol = 10^-4,
  b.start = NULL,
  d.start = NULL,
  trace = T,
  max.iter = 100,
  r = NULL
)
}
\arguments{
\item{y}{response variable, must be a vector}

\item{x}{data frame of explanatory variables}

\item{tx}{spline variable, must be a vector}

\item{prop.knots}{proportion of nodes to be interpolated, if NULL the value will be estimated through cross validation}

\item{intercept}{if model has to be estimated with an intercept}

\item{n.basis}{number of basis for estimating the B-spline}

\item{alpha}{tuning parameter for the bending of the spline, if NULL the value will be estimated through cross validation}

\item{tol}{tolerance parameter}

\item{b.start}{initial values for beta parameters}

\item{d.start}{initial values for delta parameters}

\item{trace}{if TRUE, information is printed during the running of optimization process}

\item{max.iter}{maximum number of iterations}

\item{r}{sequence of values form which the basis will be estimated. It mainly has an internal use, it's advisable to not change it}
}
\value{
PartialSplines return an object of class "PSM" which is a list containing several objects
\itemize{
  \item beta - optimal estimates of linear parameters
  \item delta - optimal estimates of spline parameters
  \item call - call of the function
  \item Z - matrix Z of spline coefficients
  \item y - response variable
  \item X - model matrix of linear variables
  \item tx - spline variable
  \item data - data frame of input explanatory varibales
  \item r - vector of points in which basis functions were calculated
  \item K - matrix used in backfitting
  \item time - execution time
  \item iter - number of iteration until convergence
  \item alpha - value of alpha used
  \item prop.knots - proportion of nodes interpolated
  \item details - list containing other elements for internal use

}
}
\description{
PartialSplines function allows you to estimate a model where the linear relationship between the response variable and the explanatory ones is summarized by coefficients Beta; on the other hand, coefficients Delta explain the unspecified functional relationship between y and the spline variable 'tx'.
}
\examples{
x <- data.frame(x = rnorm(100))
tx <- rnorm(100)
eps <- rnorm(100)
y <- x[,1] + tx^4 + eps
psm <- PartialSplines(y, x, tx)

}
