% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.PSM.R
\name{plot.PSM}
\alias{plot.PSM}
\title{Plot for a PSM object}
\usage{
\method{plot}{PSM}(object, legend.pos = "topleft", ...)
}
\arguments{
\item{object}{an object of class 'PSM'}

\item{legend.pos}{position of the legend. Can be moved according to points position}

\item{...}{further arguments passed to the generic 'plot' function}
}
\value{

}
\description{
There are two plots: in the first one there are the residuals while the second one shows the observed against fitted values
}
\examples{
x <- data.frame(x = rnorm(100))
tx <- rnorm(100)
eps <- rnorm(100)
y <- x[,1] + tx^4 + eps
psm <- PartialSplines(y, x, tx)
plot(psm)
}
