% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PSM.R
\name{print.PSM}
\alias{print.PSM}
\title{Print function for a Partial Splines model}
\usage{
\method{print}{PSM}(object, ...)
}
\arguments{
\item{object}{an object of class 'PSM'}

\item{...}{further arguments passed to the generic function 'print'}
}
\value{

}
\description{
Print function of an object of class 'PSM' will give you the main details of the model such as linear coefficients and their significance t-test, delta coefficients and a few more things
}
\examples{

x <- data.frame(x = rnorm(100))
tx <- rnorm(100)
eps <- rnorm(100)
y <- x[,1] + tx^4 + eps
psm <- PartialSplines(y, x, tx)
print(psm)
}
